Import exFat driver
This exFAT filesystem module for Linux kernel is based on sdFAT drivers by Samsung, which is used with their smartphone lineups.
The main objective of exfat-linux is to provide the best generic kernel drivers for exFAT. That means Samsung-specific modifications such as fat12/16/32 handlings, defrag and etc has been removed to make the code portable.

Maintainer of exfat: arter97
https://github.com/arter97/exfat-linux